feat: #231 forget action
