+++ crates/oxvg_actions/src/actions.rs
@@ -1,24 +1,14 @@
-use oxvg_ast::{
-    arena::Allocator,
-    node::Ref,
-    serialize::{Node, PrinterOptions, ToValue},
-};
-use oxvg_collections::{
-    atom::Atom,
-    attribute::{
-        core_attrs::Integer,
-        list_of::{ListOf, SpaceOrComma},
-    },
-};
+use oxvg_ast::{arena::Allocator, node::Ref, serialize::Node};
+use oxvg_collections::atom::Atom;
 
-use oxvg_parse::Parse;
 #[cfg(feature = "wasm")]
 use tsify::Tsify;
 
+mod state;
+
 use crate::{
     error::Error,
-    state::{DerivedState, State, StateElement},
-    utils::create_oxvg_attr,
+    state::{DerivedState, State},
 };
 
 /// An actor holds a reference to a document to act upon.
@@ -38,6 +28,8 @@ pub struct Actor<'input, 'arena> {
 #[derive(Debug, Clone)]
 /// An action is a method that an actor can execute upon a document
 pub enum Action<'input> {
+    /// See [`Actor::forget`]
+    Forget,
     /// See [`Actor::select`]
     Select(Atom<'input>),
 }
@@ -46,6 +38,8 @@ pub enum Action<'input> {
 #[napi]
 /// An action is a method that an actor can execute upon a document
 pub enum ActionNapi {
+    /// See [`Actor::forget`]
+    Forget,
     /// See [`Actor::select`]
     Select(String),
 }
@@ -96,97 +90,9 @@ impl<'input, 'arena> Actor<'input, 'arena> {
     /// When the associated action fails
     pub fn dispatch(&mut self, action: Action<'input>) -> Result<(), Error<'input>> {
         match action {
-            Action::Select(query) => self.select(query.as_str()),
+            Action::Forget => self.forget(),
+            Action::Select(query) => return self.select(query.as_str()),
         }
-    }
-
-    /// Updates the state of the actor to point to the elements matching the given selector.
-    /// Elements can also be selected by a space/comma separated list of allocation-id
-    /// integers.
-    ///
-    /// # Errors
-    ///
-    /// When root element is missing or the query cannot be parsed.
-    ///
-    /// # Spec
-    ///
-    #[doc = include_str!("./spec/select.md")]
-    pub fn select(&mut self, query: &str) -> Result<(), Error<'input>> {
-        let Some(root) = self.root.element() else {
-            return Err(Error::NoRootElement);
-        };
-        self.state
-            .record(&Action::Select(query.to_string().into()), &self.allocator);
-
-        let selections: ListOf<Integer, SpaceOrComma> =
-            if query.chars().next().is_some_and(|c| c.is_ascii_digit()) {
-                ListOf::parse_string(query).map_err(|err| Error::ParseError(err.to_string()))?
-            } else {
-                let elements = root
-                    .select(query)
-                    .map_err(|_| Error::InvalidSelector(query.to_string()))?;
-
-                #[allow(clippy::cast_possible_wrap)]
-                let selections: Vec<_> = elements.map(|e| e.id() as Integer).collect();
-
-                ListOf {
-                    list: selections,
-                    separator: SpaceOrComma,
-                }
-            };
-        self.state
-            .get_selections(&self.allocator)
-            .set_attribute(create_oxvg_attr(
-                StateElement::SELECTION_IDS,
-                selections
-                    .to_value_string(PrinterOptions::default())
-                    .map_err(|err| Error::SerializeError(err.to_string()))?
-                    .into(),
-            ));
-        self.state.embed(self.root)?;
         Ok(())
     }
-}
-
-#[test]
-fn select_empty() {
-    oxvg_ast::parse::roxmltree::parse(
-        r#"<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" width="100" height="100"/>"#,
-        |root, allocator| {
-            let mut actor = Actor::new(root, allocator).unwrap();
-
-            actor.select("svg").unwrap();
-            insta::assert_snapshot!(actor.root.serialize().unwrap());
-
-            actor.select("1").unwrap();
-            insta::assert_snapshot!(actor.root.serialize().unwrap());
-        },
-    )
-    .unwrap();
-}
-
-#[test]
-fn select() {
-    oxvg_ast::parse::roxmltree::parse(
-        r#"<svg xmlns="http://www.w3.org/2000/svg">
-    <g color="black"/>
-    <g color="BLACK"/>
-    <path fill="rgb(64 64 64)"/>
-    <path fill="rgb(64, 64, 64)"/>
-    <path fill="rgb(86.27451%,86.666667%,87.058824%)"/>
-    <path fill="rgb(-255,100,500)"/>
-</svg>"#,
-        |root, allocator| {
-            let mut actor = Actor::new(root, allocator).unwrap();
-
-            actor.select("path").unwrap();
-            insta::assert_snapshot!(actor.root.serialize().unwrap());
-            insta::assert_debug_snapshot!(actor.derive_state().unwrap());
-
-            actor.select("7, 9").unwrap();
-            insta::assert_snapshot!(actor.root.serialize().unwrap());
-            insta::assert_debug_snapshot!(actor.derive_state().unwrap());
-        },
-    )
-    .unwrap();
 }
