+++ crates/oxvg_actions/src/spec/state/forget.md
@@ -1,0 +1,9 @@
+> [!CAUTION]
+> This will temporarily remove OXVG's history, and should only be used for finalising SVGs.
+
+Removes `oxvg:state` from the document, thereby temporarily losing any associated state.
+
+```sh
+# Effects: History, Selection
+-forget
+```
